server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
aiofiles>=24.0.0
Pillow>=10.4.0
pyyaml>=6.0.2
requests>=2.32.0
